414 add new opc ua ctrl commands

+++ src/OpcUaCtrl/CaCertCtrlCommand.cpp
@@ -70,6 +70,7 @@ namespace OpcUaCtrl
 		std::cout << "        add a new ca certificate" << std::endl;
 		std::cout << "    del <Application-Name> <Cert-Id>:" << std::endl;
 		std::cout << "        delete a ca certificate" << std::endl;
+		std::cout << "        use *all* as certId to delete all ca certs" << std::endl;
 		std::cout << "    trust <Application-Name> <Cert-Id>:" << std::endl;
 		std::cout << "        activate a ca certificate" << std::endl;
 		std::cout << "    untrust <Application-Name> <CertId>:" << std::endl;
@@ -124,12 +125,24 @@ namespace OpcUaCtrl
 		// read all files from ca trust directory
 		boost::filesystem::path caDirectoryTrust(applicationInfo->caDirectoryTrust_);
 		for ( auto filenameIt : boost::filesystem::directory_iterator(caDirectoryTrust) ) {
+			// show only ca certificates
+			if (!isCaRootCertificate(filenameIt.path().string())) {
+				continue;
+			}
+
+			// show certificate info
 			showCertificateInfo(ss, filenameIt.path().string(), "Trusted");
 		}
 
 		// read all files from ca revocation directory
 		boost::filesystem::path caDirectoryRevocation(applicationInfo->caDirectoryRevocation_);
 		for ( auto filenameIt : boost::filesystem::directory_iterator(caDirectoryRevocation) ) {
+			// show only ca certificates
+			if (!isCaRootCertificate(filenameIt.path().string())) {
+				continue;
+			}
+
+			// show certificate info
 			showCertificateInfo(ss, filenameIt.path().string(), "Untrusted");
 		}
 
@@ -152,11 +165,6 @@ namespace OpcUaCtrl
 		Certificate certificate;
 		if (!certificate.fromDERFile(filename)) {
 			certificate.log(Error, "read certificate in der format error");
-			return false;
-		}
-
-		// check whether it is a ca certificate
-		if (!certificate.isCaCertificate()) {
 			return false;
 		}
 
@@ -218,8 +226,8 @@ namespace OpcUaCtrl
 		}
 
 		// check whether it is a ca certificate
-		if (!certificate.isCaCertificate()) {
-			certificate.log(Error, "ca certificate error");
+		if (!certificate.isCaRoot()) {
+			certificate.log(Error, "root ca certificate error");
 			return 1;
 		}
 
@@ -277,6 +285,32 @@ namespace OpcUaCtrl
 		}
 		auto applicationInfo = it->second;
 
+		// check whether all certificates should be deleted
+		if (certificateId == "all") {
+
+			// find and delete all ca certificates in trusted directory
+			for (auto& entry : boost::filesystem::directory_iterator(applicationInfo->caDirectoryTrust_)) {
+				std::vector<std::string> cmdLine = commandLine;
+				cmdLine[4] = entry.path().stem().string();
+				auto rc = del(cmdLine);
+				if (rc != 0) {
+					return rc;
+				}
+			}
+
+			// find and delete all ca certificates in revocation list directory
+			for (auto& entry : boost::filesystem::directory_iterator(applicationInfo->caDirectoryRevocation_)) {
+				std::vector<std::string> cmdLine = commandLine;
+				cmdLine[4] = entry.path().stem().string();
+				auto rc = del(cmdLine);
+				if (rc != 0) {
+					return rc;
+				}
+			}
+
+			return 0;
+		}
+
 		// generate path and name for trusted and revocation ca certificate file
 		boost::filesystem::path targetTrustCertificateFile(
 			applicationInfo->caDirectoryTrust_ +
@@ -304,6 +338,14 @@ namespace OpcUaCtrl
 
 		// remove certificate
 		if (boost::filesystem::exists(targetTrustCertificateFile)) {
+			// remove only ca root certificates
+			if (!isCaRootCertificate(targetTrustCertificateFile.string())) {
+				Log(Error, "certificate is not a ca root certificate")
+					.parameter("ApplicationName", applicationName)
+					.parameter("CertId", certificateId);
+				return 1;
+			}
+
 			if (!boost::filesystem::remove(targetTrustCertificateFile)) {
 				Log(Error, "remove certificate error")
 					.parameter("ApplicationName", applicationName)
@@ -311,6 +353,14 @@ namespace OpcUaCtrl
 			}
 		}
 		if (boost::filesystem::exists(targetRelocationCertificateFile)) {
+			// remove only ca root certificates
+			if (!isCaRootCertificate(targetRelocationCertificateFile.string())) {
+				Log(Error, "certificate is not a ca root certificate")
+					.parameter("ApplicationName", applicationName)
+					.parameter("CertId", certificateId);
+				return 1;
+			}
+
 			if (!boost::filesystem::remove(targetRelocationCertificateFile)) {
 				Log(Error, "remove certificate error")
 					.parameter("ApplicationName", applicationName)
@@ -366,6 +416,14 @@ namespace OpcUaCtrl
 			Log(Error, "certificate identifier unknown")
 				.parameter("ApplicationName", applicationName)
 				.parameter("CertId", certificateId);
+		}
+
+		// certificate must be a ca root certificate
+		if (!isCaRootCertificate(targetRelocationCertificateFile.string())) {
+			Log(Error, "certificate is not a ca root certificate")
+				.parameter("ApplicationName", applicationName)
+				.parameter("CertId", certificateId);
+			return 1;
 		}
 
 		// trust ca certificate
@@ -428,6 +486,14 @@ namespace OpcUaCtrl
 				.parameter("CertId", certificateId);
 		}
 
+		// certificate must be a ca root certificate
+		if (!isCaRootCertificate(targetTrustCertificateFile.string())) {
+			Log(Error, "certificate is not a ca root certificate")
+				.parameter("ApplicationName", applicationName)
+				.parameter("CertId", certificateId);
+			return 1;
+		}
+
 		// untrust ca certificate
 		boost::system::error_code ec;
 		boost::filesystem::rename(targetTrustCertificateFile, targetRelocationCertificateFile, ec);
@@ -438,6 +504,24 @@ namespace OpcUaCtrl
 		}
 
 		return 0;
+	}
+
+	bool
+	CaCertCtrlCommand::isCaRootCertificate(const std::string& certFileName)
+	{
+		// read certificate from file
+		Certificate certificate;
+		if (!certificate.fromDERFile(certFileName)) {
+			certificate.log(Error, "read certificate in der format error");
+			return false;
+		}
+
+		// check whether it is a root ca certificate
+		if (!certificate.isCaRoot()) {
+			return false;
+		}
+
+		return true;
 	}
 
 }
